setup: Ensure the parent directory exists when extracting
It seems that when reading the zip archives from FreeDOS there are
entries such as 'bin/file.exe' which will need the parent directories
creating. Currently the extraction only creates directories if the
entry is marked as a directory. Since extraction used to work, I can
only imagine that in the past the archive would have had a preceding
directory entry called 'bin'.